Update code-rules.tex
@kayleachampion

Line 158 -159 ("the threats to your project are political as well as technological") could include a pointer back to the threat modeling tip (whatever tip number it winds up as). Much of what's said in this tip is tied to threat modeling.
Line 207 "also fosters community collaboration " -- missing '.' at the end of the sentence.